[SYCL] Disable global offset testing for a little while
Global offset support is being added to Update 1.  Previously,
this test failed because global offset support was not present.
Now this test will pass if the system has L0 ci-neo-018674 or later.
Rather than choosing between unexpectedly passing or failing
as expected, this test is disabled.

The global offset support is experimental because the needed API
was not part of the Level Zero 1.0 Specification.  Therefore, this
support cannot be added to github.

CMPLRLLVM-25796 tracks the updating of this test.

Signed-off-by: Gail Lyons <gail.lyons@intel.com>
Change-Id: I196aba16ed5da46c79c4117c1af15fb66dc0de5a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/309337
Tested-by: CE Jenkins <cejenkins@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Reviewed-by: yaman, wang <yaman.wang@intel.com>