[SYCL][ESIMD] Provide the for_all_combinations utility (#721)
[SYCL][ESIMD] Provide the for_all_combinations utility

Provides a generic way to represent a combinations of types and compile-time
values to cover. Supports any number of type and value packs to iterate over

Adds unnamed type pack support

By doing so we avoid additional restriction on architecture of
the test case functors and on logging architecture, as we don't
need to construct string description for every type/value on the
top level.
We are still free to use named type packs, giving us a possibility
to have any combination of named packs and unnamed packs, with value
packs supported in both variations.
Tests are updated as there is no more for_all_types_and_dims function

Signed-off-by: Kochetkov, Yuriy <yuriyx.kochetkov@intel.com>