[SYCL][ESIMD] Make logs architecture more flexible (#838)
Removes tight coupling of the test case description and error message, so that
we could to use any error messages with no need of manual message construction.
Any log message specific details are moved to the upper level, to the message
construction itself, because log messages are entry points for test failure
analysys.
Also a syntax sugar for trace and debug messages added.

Signed-off-by: Kochetkov, Yuriy yuriyx.kochetkov@intel.com