[SYCL][HIP] Fix assert tests (#1083)
Assert is now implemented for HIP.

Additionally fixing the assert multiple tus tests led to fixing the
`sycl-external.cpp` test. It turns out that the AMDGPU llvm backend was
running some dead code elimination and deleting `SYCL_EXTERNAL`
functions, this is also why `kernel_bundle_ignore_sycl_external.cpp`
worked on AMD, not because it was behaving correctly but because the DCE
was deleting the `SYCL_EXTERNAL` function so it looked "ignored" from
the kernel bundle, this test is now marked as failing again.